+++ demo/startWorker.php
@@ -1,13 +1,22 @@
 <?php
 
 use Qless\Client;
+use Qless\Jobs\Reservers\OrderedReserver;
+use Qless\Queue;
 use Qless\Worker;
 
 require_once __DIR__ . '/../tests/bootstrap.php';
 
-$queues = ['testQueue1', 'testQueue2'];
-
 $client = new Client(REDIS_HOST, REDIS_PORT, REDIS_TIMEOUT);
-$worker = new Worker("TestWorker_1", $queues, $client, 5);
 
+$worker = new Worker(
+    new OrderedReserver(
+        array_map(function (string $name) use ($client) {
+            return new Queue($name, $client);
+        }, ['test-queue-1', 'test-queue-2'])
+    ),
+    $client
+);
+
+$worker->setInterval(5);
 $worker->run();