+++ src/Jobs/Reservers/ShuffledRoundRobinReserver.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace  Qless\Jobs\Reservers;
+
+/**
+ * Qless\Jobs\Reservers\ShuffledRoundRobinReserver
+ *
+ * @package Qless\Jobs\Reservers
+ */
+class ShuffledRoundRobinReserver extends RoundRobinReserver
+{
+    const TYPE_DESCRIPTION = 'shuffled round robin';
+
+    /**
+     * {@inheritdoc}
+     *
+     * @return void
+     */
+    public function beforeFork(): void
+    {
+        shuffle($this->queues);
+        $this->resetDescription();
+    }
+}